refactor into two files, remove class
