+++ .github/linters/.python-lint
@@ -1,2 +1,2 @@
 [MESSAGES CONTROL]
-disable=attribute-defined-outside-init, global-statement, invalid-name, import-error, line-too-long, missing-class-docstring, missing-function-docstring, missing-module-docstring, no-member, no-name-in-module, protected-access, redefined-builtin, redefined-outer-name
+disable=duplicate-code, line-too-long, missing-function-docstring, missing-module-docstring, no-name-in-module, no-name-in-module, protected-access, redefined-outer-name
