+++ .pre-commit-config.yaml
@@ -22,11 +22,6 @@ repos:
       args: ["--config", ".github/linters/.markdown-lint.yaml"]
   - repo: local
     hooks:
-      - id: embedme
-        name: embedme
-        entry: npx
-        language: system
-        args: ["embedme", "--verify", "README.md"]
       - id: pylint
         name: pylint
         entry: pylint
@@ -38,7 +33,7 @@ repos:
         entry: pytest
         language: system
         types: [python]
-        args: ["test/test.py", "-W", "ignore::DeprecationWarning", "--ollama"]
+        args: ["test", "-W", "ignore::DeprecationWarning"]
       - id: cypress
         name: cypress
         entry: bash
