+++ Dockerfile
@@ -21,9 +21,7 @@ RUN ollama serve & \
 
 RUN python3 -m pip install --no-cache-dir -r requirements.txt
 
-RUN mkdir memory
-
 EXPOSE 7860
 
 ENTRYPOINT []
-CMD ["bash", "-c", "/bin/ollama serve & chainlit run main.py -h --port 7860"]
+CMD ["bash", "-c", "/bin/ollama serve & chainlit run app.py -h --port 7860"]
