+++ cypress/test.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-chainlit run main.py -h -c &
+chainlit run app.py -h -c &
 sleep 10
 npx cypress run --record false --config-file cypress/cypress.config.ts
 kill %%
